apply norminette other than expand_handler
