+++ Makefile
@@ -6,7 +6,7 @@
 #    By: tamatsuu <tamatsuu@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/10/25 21:34:48 by shokosoeno        #+#    #+#              #
-#    Updated: 2024/12/30 22:08:22 by tamatsuu         ###   ########.fr        #
+#    Updated: 2024/12/31 03:15:27 by tamatsuu         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -68,7 +68,12 @@ srcs/signal/signal_child.c \
 srcs/execution/execute_fd_controller.c \
 srcs/execution/execute_logical_operator.c \
 srcs/execution/execute_utils.c \
-srcs/execution/execute.c 
+srcs/execution/execute.c \
+srcs/expand/expand_handler.c \
+srcs/expand/expand_helper.c \
+srcs/expand/expand_variable_helper.c \
+srcs/expand/expand_variable.c \
+srcs/expand/remove_quotes.c 
 
 # srcs/lexer/lexer_helper.c \
 # srcs/lexer/lexer_token_checker.c \
