+++ includes/expand.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   expand.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tamatsuu <tamatsuu@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: tamatsuu <tamatsuu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 18:27:18 by tamatsuu          #+#    #+#             */
-/*   Updated: 2024/12/18 20:41:17 by tamatsuu         ###   ########.fr       */
+/*   Updated: 2024/12/31 02:40:17 by tamatsuu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,19 +17,21 @@
 # define UNDER_SCORE '_'
 # include "./parser.h"
 # include "./lexer.h"
-# include "./builtin.h"
+# include "./map.h"
 
-void	expand_handler(t_node *node, char **envp);
-void	expand_variable_handler(t_node *node, char **envp);
-char	*expand_variable(char *str, char **envp);
+void	expand_handler(t_node *node, t_map *envp);
+void	expand_variable_handler(t_node *node, t_map *envp);
+char	*expand_variable(char *str, t_map *envp);
 size_t	retrieve_val_in_sq(char **ret, char *str, size_t i);
 size_t	skip_s_quote(char *str, size_t i);
 bool	is_dollar_symbol(char c);
-size_t	retrieve_val_in_dq(char **ret, char *str, size_t i, char **envp);
+size_t	retrieve_val_in_dq(char **ret, char *str, size_t i, t_map *envp);
 char	*str_concat_helper(char **ret, char *str, size_t i, size_t len);
-size_t	retrieve_var(char **ret, char *str, size_t i, char **envp);
+size_t	retrieve_var(char **ret, char *str, size_t i, t_map *envp);
 size_t	retrieve_var_name_len(char *str, size_t i);
-t_map	*init_envmap(char **envp);
 size_t	retrieve_normal_val(char **ret, char *str, size_t i);
+void	remove_quote_handler(t_node *node);
+char	*remove_quotes(char *str);
+int		move_to_next_quotation_expnd(char *input, int i);
 
 #endif
