+++ test/expand_test/Makefile
@@ -6,7 +6,7 @@
 #    By: tamatsuu <tamatsuu@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/10/25 21:34:48 by shokosoeno        #+#    #+#              #
-#    Updated: 2024/12/16 23:15:25 by tamatsuu         ###   ########.fr        #
+#    Updated: 2024/12/31 03:25:28 by tamatsuu         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -35,8 +35,20 @@ SRCS = ./expand_tester.c \
 ../../srcs/map/map_set.c \
 ../../srcs/map/item_utils.c \
 ../../srcs/utils/error.c \
-../../srcs/expand/expand.c \
-../../srcs/lexer/lexer_token_quote_checker.c
+../../srcs/utils/xmalloc.c \
+../../srcs/expand/expand_handler.c \
+../../srcs/expand/expand_helper.c \
+../../srcs/expand/expand_variable_helper.c \
+../../srcs/expand/expand_variable.c \
+../../srcs/expand/remove_quotes.c \
+../../srcs/lexer/lexer_helper.c \
+../../srcs/lexer/lexer_token_checker.c \
+../../srcs/lexer/lexer_token_quote_checker.c \
+../../srcs/lexer/lexer_operator_checker.c \
+../../srcs/lexer/lexer.c \
+../../srcs/utils/init_env.c \
+../../srcs/map/map_size.c \
+../../srcs/map/map_unset.c 
 
 
 OBJS = $(SRCS:.c=.o)
