+++ test/expand_test/expand_tester.c
@@ -3,37 +3,43 @@
 /*                                                        :::      ::::::::   */
 /*   expand_tester.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tamatsuu <tamatsuu@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: tamatsuu <tamatsuu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 22:55:35 by tamatsuu          #+#    #+#             */
-/*   Updated: 2024/12/18 01:36:56 by tamatsuu         ###   ########.fr       */
+/*   Updated: 2024/12/31 03:22:27 by tamatsuu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "./expand_tester.h"
+#include <string.h>
+#include <stdlib.h>
 
 int	main(int argc, char *argv[], char *envp[])
 {
 	char	*ret;
+	t_map	*env;
 
 	ret = NULL;
+	env = init_env(envp);
 	if (!argv || argc < 0)
 		return (1);
 
 	if (argv[1])
 	{
-		ret = expand_variable(argv[1], envp);
+		ret = expand_variable(argv[1], env);
 		printf("result %s\n",ret);
 	}
 	else
-		test_expand_variable(envp);
+		test_expand_variable(env);
 }
 
-void	test_expand_variable(char *envp[])
+void	test_expand_variable(t_map *envp)
 {
 	char	*ret;
+	char	*after_ret;
 
 	ret = NULL;
+	after_ret = NULL;
 	char	*test_cases[] ={
 		"$USER",
 		"'$USER'",
@@ -57,7 +63,14 @@ void	test_expand_variable(char *envp[])
 	{
 		ret = expand_variable(test_cases[i], envp);
 		printf("result %s\n",ret);
+		if (strchr(ret, '\'') || strchr(ret, '\"'))
+		{
+			after_ret = remove_quotes(ret);
+			printf("result after remove quote: %s\n", after_ret);	
+		}
 		free(ret);
+		free(after_ret);
 		ret = NULL;
+		after_ret = NULL;
 	}
 }
