+++ test/expand_test/expand_tester.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   expand_tester.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tamatsuu <tamatsuu@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: tamatsuu <tamatsuu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 22:54:28 by tamatsuu          #+#    #+#             */
-/*   Updated: 2024/12/18 00:00:23 by tamatsuu         ###   ########.fr       */
+/*   Updated: 2024/12/31 03:20:55 by tamatsuu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,8 @@
 # define EXPAND_TESTER_H
 
 # include "../../includes/expand.h"
-void	test_expand_variable(char *envp[]);
+# include "../../includes/environment.h"
+void	test_expand_variable(t_map *envp);
 
 
 #endif
